A few string.h changes.
* Changed strlen to return a size_t.
* Changed vsprintf and puts accordingly.
* Added strcpy and strncpy.

+++ src/kernel/vsprintf.c
@@ -164,7 +164,7 @@ int vsprintf(char *buf, const char *fmt, va_list args)
 
 		case 's':
 			s = va_arg(args, char *);
-			len = strlen(s);
+			len = (int) strlen(s);
 			if (precision < 0)
 				precision = len;
 			else if (len > precision)
